style stations-show to include get historic streamflow button and js to fetch streamflow from api endpoint
