+++ app/assets/stylesheets/stations-show.scss
@@ -1,10 +1,8 @@
 .station-show-info {
-  // display: flex;
-  // justify-content: center;
-  // align-items: center;
   height: 200px;
   width: 300px;
   background-color: rgb(131, 159, 159);
+  margin-top: 50px;
 }
 
 .mid-nav-2 {
@@ -18,14 +16,57 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  position: relative;
-  top: 90px;
-  right: 75%;
   background-color: rgb(181, 213, 183);
   width: 740px;
   height: auto;
-  margin-bottom: 50px;
+  margin-top: 50px;
   .chart-title {
     margin-bottom: 10px;
   }
+}
+
+.historical-discharge-search {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  background-color: rgb(171, 191, 183);
+  z-index: 1;
+  min-width: 300px;
+  width: 350px;
+  border: 2px solid black;
+  border-radius: 2px;
+  padding: 15px;
+  margin-bottom: 50px;
+  margin-top: 50px;
+  h3 {
+    text-align: center;
+    margin-top: 0px;
+  }
+}
+
+.search-container {
+  display: flex;
+  align-items: center;
+  width: auto;
+}
+
+.get-historic-container {
+  display: flex;
+  justify-content: center;
+}
+
+.get-historic {
+  margin-top: 10px;
+  width: 60%;
+  font-size: 12px;
+  font-weight: bold;
+  background-color: rgb(169, 185, 153);
+}
+
+.email-streamflow {
+  margin-top: 10px;
+  width: 55%;
+  font-size: 12px;
+  font-weight: bold;
+  background-color: rgb(177, 201, 224);
 }
