+++ app/services/usgs_streamflow_service.rb
@@ -1,5 +1,4 @@
 class USGSStreamflowService
-  include USGSJsonResponse
 
   def initialize(search_params)
     @search_params = search_params
@@ -14,9 +13,23 @@ class USGSStreamflowService
     attr_reader :search_params
 
     def base_params
-      {
-        "format" => "json",
-        "period" => "P1D"
-      }
+      {"format" => "json"}
     end
+
+    def conn
+      Faraday.new 'http://waterservices.usgs.gov' do |conn|
+        conn.use FaradayMiddleware::FollowRedirects, limit: 5
+        conn.adapter Faraday.default_adapter
+      end
+    end
+
+    def send_get_request(uri, params)
+      conn.get(uri, params)
+    end
+
+    def json_response(uri, params)
+      response = send_get_request(uri, params)
+      JSON.parse(response.body)
+    end
+
 end
